functions/requirements.txt
azure-functions
requests
azure-storage-blob
azure-identity